Process refcnt optimizations
The __startcore optimization is similar to how it was before, but I
separated it from the owning_proc commit to show that it isn't necessary
and to make it more clear.

The kthread optimization is just a comment for now, since it's not clear
to me which way is best (if any).