Kthreads from a different process can restart
Whether or not you want to run kthread bottom halves on a core that is
owned by another process is a matter of policy.  The code can handle it
either way.  If you don't want to, then route your interrupts to LL
cores and put some thought into where you send the __launch_kthread
KMSG.