DONT_MIGRATE set more carefully
It's held the entire time throughout an mcs_notif_safe lock, since it
doesn't make sense to migrate or otherwise try to restart/run that
thread (could be holding the lock, etc).

Also, after turning off DONT_MIGRATE, you need to check messages (which
enable_notifs() does).  This isn't critical right now, but future
patches will require it.  It's harmless to turn them on before
enable_notif(), since the protection it provides is in the process of
disabling notifs.  Once that is done, you won't get migrated anyways.

Finally, this makes sure we only do the DONT_MIGRATE / notif_disabling
business if we're in _M mode.  This wasn't pairing properly, and we were
getting our thread0 set with DONT_MIGRATE the first time it ran.