proc_notify() no longer checks notif_pending
We were accidentally masking our IPIs when notif_pending was set.  This
was due to a change over time of what notif_pending means.  It used to
only mean "an IPI is on its way", but now it means "there is a reason to
reenter vcore context and/or check messages/events".