Quickly return/pop DONT_MIGRATE uthreads
Two rules:

1) Don't check messages/handle events when you have a DONT_MIGRATE
uthread.  This will become important in the future.

2) All uses of DONT_MIGRATE must reenable notifs (and check messages) at
some point.  One such case is uthread_yield().  Another is
mcs_unlock_notifsafe().

This patch handles the case where a uthread turns on DONT_MIGRATE and
gets IPI'd before it can disable_notif.  We simply return to it, leaving
notif_pending turned on so that we make sure we will go back into vcore
context and check messages at the next safe opportunity.  That makes
this the one time we can leave vcore context with notif_pending: since
we know we'll get back to it right away.

Another way to look at it is that we lost the race, and got the IPI
before disabling.  This allows us to pretend like we won the race
(barring any bugs...).