Uthread helpers for disabling notifs
The DONT_MIGRATE bit is tricky, and that code is about to get repeated a
lot, so it makes sense to have a helper.  Other code might be able to
use this, though I left them as is.  uthread_init() seems clear as is,
and might be getting mucked with in a concurrent patch set.