upgraded to gcc 4.6.1 and glibc 2.14.1 (XCC)
This commit includes not only brigning in the new gcc and glibc, but also the
changes to the glibc-*-ros files as well.  We did this so that this commit
will compile and build properly.  To see the diff of the glibc-*-ros files
with the old version, run the following command relative to checking out THIS
commit:

git diff HEAD^:tools/compilers/gcc-glibc/glibc-2.11.1-ros \
  HEAD:tools/compilers/gcc-glibc/glibc-2.14.1-ros