Alarm interface for incremental alarms
So you can reset an alarm for a new time, relative to the last alarm
(compared to 'relative to now').  This is useful if you want a periodic
tick, where you care about how much time passes.

Though you can always run into issues where you try to tick too fast for
the system to keep up, and you'll start to fall behind (setting alarms
for the past, etc), so anything serious is likely to need a more serious
solution.