x86 console and readline() honor 0x7f
Qemu sends an 0x7f (DEL) via serial when you hit backspace, and readline
was ignoring it.  Not that readline matters much, but it'll now treat
the 0x7f like a backspace (\b, 0x8).

Note that \b and \x7f get sent directly to userspace, and it is up to
them to parse that.  This means that the KB (qemu and real world)
backspace is sending \b, real-world serial (at least my minicom) is
sending \b, but qemu serial sends \x7f.

Keymaps, terminal settings, and what programs expect are all huge pains
in the ass.