Fake tcgetattr() to allow busybox to do line edits
This fakes the tcgetattr() call with hard-coded answers from a linux
machine.  This is enough to get busybox to do line editing.  It also
allows us to remove the devfs stdin hacks, and add new ones to stdout!

I added my busybox config to tools/.  If you don't want to use mine, be
sure to turn on FEATURE_EDITING in yours to get backspace working on the
command line.  I'll update it periodically.