Pthread state tracking
We don't do anything smart in pth_thread_runnable(), but that's where
you can make different decisions on what to do with the thread based on
why it slept/woke up.

All yielding/blocking paths will touch the 2LS (pthread in this case),
and that is where the 2LS can record info (states) for later
(thread_runnable).