proc_destroy() now starts in the ksched
This allows the kernel scheduler to lock first, which helps with some
other lock-ordering trickery we do, and allows the ksched to do
additional cleanup, other than just receiving the idle cores.

Additionally, we have less of a need for put_idle_cores() (the only
caller is an old preemption function, which might get sucked into the
ksched soon).