Tracks state properly for paused uthreads
Pthreads uses a state for threads stopped from below (currently only for
preemptions (either pending or recoverery)).

Uthreads just tracks them as NOT_RUNNABLE.

This patch also provides a helper for the three usages of
copyout/sched_ops->paused.  Depending on what 2LSs need, we might
provide copyout directly again, but for now we'll leave it as an
internal helper.