proc_change_to_m() is now a scheduler function
Due to issues of lock ordering.  We need to atomically change process
state and tell the ksched about it.  The way to do this without
deadlocking is to grab the ksched lock first.  Ultimately, we're going
to end up tying proc and ksched code together a bit, where proc code
does the bidding of sched code.