proc_wakeup() - spammable and starts in the ksched
Need to have the ksched lock held while waking up so that we atomically
wake up and alert the ksched (via the callbacks).  If we grab the ksched
lock after the proc lock, we can deadlock.  If we unlock the proc lock
before telling the ksched, we could have another core try a
proc_destroy(), and the ksched would destroy it after the old
__proc_wakeup, but before the ksched finds out about the wakeup.  So we
enter the call via the ksched, and grab its lock first, like with
destroy().

Note that excessive spamming of proc_wakeup would hurt the ksched a bit
(coarser lock than all the proc locks), but all cases expect it to be a
necessary call.  For instance, the event code all peaks at the state of
p, making sure it was WAITING before bothering the ksched.  We'd have
excessive calls only if two or more cores tried that at the same time
and all saw WAITING before the ksched called __proc_wakeup to change it
to RUNNABLE.