proc_wakeup() replaces schedule_scp()
All places wanting to make sure a proc runs now can spam proc_wakeup().
This commit makes new-proc calls and yield-related calls that aren't
already in the ksched call proc_wakeup().

This should also have cleaned up the last of the deadlockable lock
orderings (ever since we had proc_destroy in the ksched, it was possible
to deadlock during a concurrent sched_scp()).

Finally, this just gets rid of schedule_scp(), which was in no way
related to the __schedule_scp().  Ugh.