Ksched functions to provision cores
You can provision cores, and this code will track where each core is
provisioned and/or allocated to.  The scheduler does nothing with this
info yet.

Also some CG/LL changes - notably we don't reserve a core for the NIC
anymore, and if you use any of the other antiquated core reservation
CONFIGs, something will break when you provision that core.  That'll get
sorted out eventually.