Ksched allocates provisioned cores
When servicing a proc (via the ksched's linear scan of all MCPs or via a
poke_ksched()), the ksched will preempt cores provisioned to a process
that are allocated to another process.

We also do a somewhat-ghetto trick when the ksched is attempting to
preempt a core, but is racing with that proc yielding the core.  We
short-circuit put_idle, since the ksched noticed the core is not
allocated before proc code told the ksched.