Spinlock depth checking
This tracks how many locks we acquired on a pcpu basis.  Before we
kthread/block or anything like that, the number better be 0.  Not only
is it bad for perf, but you could deadlock.

Needed this to hunt down a deadlock bug (asserted in some place where I
thought I had no locks), though it had nothing to do with sleeping.

It only does work when CONFIG_SPINLOCK_DEBUG is on, though you can
always test for it (and you'll get 0).

I didn't compile it for RISCV.  YMMV.