Fixes up amt_wanted/amt_granted in __core_request
I considered holding the proclock the entire time, and it's actually
safe to do so (we make sure we don't preempt the proc we are trying to
give the core to, which would DL).  But this way, we don't sit on the
lock forever, and only do it when we need to.  Now we're just a bit more
clear about things.  (It wasn't clear before, and was a potential issue
when hunting a brutal bug).

Oh, and this was not fixing up amt_wanted like it thought it was.