Post-and-poke style sync for the ksched
Might change some of the details of the "poke-style", but I definitely
want at least a generic version of it (and all of its mb()s) to be
kernel-wide.

This is the same style of sync used in vcore_request(), though that one
is built in directly.

Note that we still use the ksched lock within in poke func - all this
commit does is create the poke style and use it to trigger the sched
mcp.  I'll deal with the lock holding shortly.