__core_request no longer locks and calls proc code
Part of a move to make the ksched lock not lock-ordered before the proc
lock.

__core_req needs to be careful about how things change when it unlocks.
for instance, there might be callbacks for cores yielding or procs
dying, or someone might have called provision_core().  I think I've
covered the cases, but it's a bit fragile.

Also note that I'm relying on __core_req being called once at a time
only (specifically, that there are no other preemptors or allocators).
The poke_tracker gives me that.  (That, and that __core_req is only
called from within the poke).