Debug code to see remote kmsgs
Mostly useful for resolving deadlocks - you can see what kmsgs are on
remote cores.  It won't deal with races (though I can build that in if
we want).

The easiest way to use it on x86 is to "trace coretf coreid", since I
built it in the NMI handler.  Or you can kfunc print_kmsgs 0xcoreid
(remember, kfunc deals with hex!).