Ensures IRQs are enabled when proc_destroy()ing
The main reason is that irqs need to be on when grabbing the proclock.
There are some deadlock scenarios where a __preempt (or even __death)
kmsg is sent with the lock held to a core, but if that core has irqs
disabled, it won't process the message (circular waiting, etc).

While I could make proc_destroy() enable irqs, its really up to the
calling code to do that.  I put in asserts where they shouldn't be
disabled, and otherwise turned them on manually.  This mostly applies to
trap handling code.  Note we can't just turn on IRQs immediately - each
case needs to do it where it is possible (preferably sooner than later),
such as after reading cr2() in x86's PF handler.