Proc code locks before disabling IRQs
While we cannot grab the lock with interrupts off, we *can* disable
after grabbing the lock, but only under certain circumstances.  This is
because we send the kmsgs with the lock held.  Once we hold the lock, we
know no one will send proc mgmt kmsgs except us.  And it is really
important that we don't send any in these functions.  (imagine trying to
do a __preempt, when you send something to your own core and then try to
__map_vcore, and never get the message).

Also note that we must use immediate messages (which we've been doing
for a while now).  If we didn't use immediate kmsgs for __preempt,
someone (for instance) yielding or change_to-ing would spin on the lock
til they got the lock, but would never get the message since the lock
holder was waiting on them to do the __preempt.

Btw, in case it wasn't clear, the lock holder is spinning in
__map_vcore(), which is where we wait on the completion of the __preempt
kmsg.