RISCV uses the no-CAS version of MCS-PDR locks
These ones don't know how to figure out what vcore in particular needs to run,
so they make all (preempted) vcores run.  It'll eventually work out.

Also, the reason we needed to do this was because RISCV's CAS is busted and
does not allow independent access of the variable.  All accesses to (current)
RISCV CAS variables needs to be through CAS.

Note this means that registering event queues for blocking syscalls is busted.