Fixes list management bug in the ksched
A poorly timed death callback was corrupting lists by removing the dying
proc from the primary list twice.  (If death happened while the dying
proc was being serviced by __run_mcp_ksched()).

Also gets rid of the "it must be on a list, but i don't know which one"
style of list removal.