Fixes preemption handling bug
If a vcore comes up while another vcore is stealing its uthread, it
needs to check for preemptions.  There are a bunch of ways to do this.
The old way was wrong, since it assumed the preempt_ev_q had an actual
mbox, which it wasn't.  Given our current preempt_ev_q style, we want to
check our public mbox for any preempt-msg-spam, which we'd get if we
were the last core around.

Note this might not be ideal - we still have a lot of messages being
sent, and we might be able to cut that down by using a different style.
For now, it fixes the bug.