Fixes bug where syscalls were completed twice
When proc_yield() failed, the syscall would be completed twice.  When
that vcore was preempted and restarted on another core (due to a
change_to, for instance), after the first completion, the vcore would
return up the stack.  When the second completion hits, the kernel would
clobber a small part of the stack (syscalls are usually on the user
stack), which would have disasterous effects (looked like registers were
getting clobbered).