Fixes set_tls_desc() bug (XCC)
set_tls_desc() has historically set __vcoreid within the TLS to whatever
vcoreid it is being loaded on.  This was an ugly leftover from back in
the day.  It is necessary to do this for uthreads that are starting up
on a vcore (so they know which vcore it is), but we *never* want to do
this for vcore contexts.

This surfaced when handling_vc_preempt(), when we changed into a
remote vcore's TLS.  That would clobber their __vcoreid, which would lead
to all sorts of confusion (using the wrong MCS lock qnode, loading other
vcore TLSs instead of your own, telling uthread's the wrong vcoreid,
etc).

You need to rebuild glibc (copy over the edited file before building).