Fixes change_to_vcore failure case
sys_change_vcore() could fail for a few reasons.  Previously, userspace
was assuming that if it failed it was because the vcore was already
running.  However, it can fail because the caller is getting preempted.

Changes the syscall to pass back a return value based on the type of
failure.  Userspace now continues to attempt to change until it not
longer is told to try again.

Note that it is not necessary to loop in code like ensure_vcore_runs,
since those functions are already called in busy loops.