Fixes bug in generic_dir_read()
And cleans up error handling in elf.c.  This bug showed up when trying
to load "/" as an elf (telling ash "./", which thought it was a binary).
generic_dir_read() would clobber the stack of load_one_elf() due to bad
boundary checks in its for loop.

Also, makes elf failures easier to debug, and tries to catch more issues
based on read()'s retvals.