Fixes sys_change_to memory clobber
This was previously fixed in the kernel in 67e1cce, and then screwed up
recently in d9513a4.

Fixing it in the kernel is a bad way to go, so this fixes it in
userspace.  The core issue is userspace asks the kernel to do something
to basically 'free' memory (reset the VC ctx to the top of stack), and
that memory is being used by the kernel to report the syscall.  It's
somewhat equivalent to mmapping a page, putting a struct syscall on it,
then using that to request munmapping that page, and then being
surprised when something bad happens to you (we actually don't handle
this well yet).