Vcore versioning for __preempt / __startcore (XCC)
Replaces preempt_served with a better mechanism to detect preemptions.
Instead of spinning while holding the proc_lock (in __map_vcore()), we
spin in the __startcore (or __set_curtf()) handlers from KMSG context.
This sets up a happens-before ordering where we wait to make sure a
preempt happened before starting a core.

The old code can use the vcoremap, but one of the issues is that
syscalls can run for a vcore/cur_tf after that vcore has been unmapped
and then remapped again.  Simply detecting this isn't enough btw, since
it can be a real pain to deal with (imagine a page fault, or a trap of
some sort).  Better documentation to follow.

Reinstall your kernel headers and rebuild your parlib/userspace.