No longer disables irqs when mucking with pcpui
For things like cur_proc, owning_{proc,vcoreid}, and cur_tf, we don't
disable interrupts when mucking with them in proc code.  We used to need
this when proc mgmt KMSGs were immediate.  So long as we continue to not
muck with that state from IRQ handlers or IMMED KMSGs, we're okay.